+++ src/info.rs
@@ -1,5 +1,6 @@
+use crate::{Contact, License, MsCodeGenerationSetting};
+use indexmap::IndexMap;
 use serde::{Deserialize, Serialize};
-use crate::{License, Contact};
 
 /// General information about the API.
 /// https://github.com/OAI/OpenAPI-Specification/blob/master/versions/2.0.md#info-object
@@ -20,4 +21,10 @@ pub struct Info {
     pub license: Option<License>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub version: Option<String>,
+
+    /// enables passing code generation settings via OpenAPI definition
+    /// (deprecated! Please use configuration files instead.)
+    /// https://github.com/Azure/autorest/blob/master/docs/extensions/readme.md#x-ms-code-generation-settings
+    #[serde(rename = "x-ms-code-generation-settings", default, skip_serializing_if = "IndexMap::is_empty")]
+    pub x_ms_code_generation_settings: IndexMap<String, MsCodeGenerationSetting>,
 }