+++ src/lib.rs
@@ -14,18 +14,7 @@ mod security;
 mod tag;
 mod util;
 
-pub use self::autorest::*;
-pub use self::contact::*;
-pub use self::external_documentation::*;
-pub use self::header::*;
-pub use self::info::*;
-pub use self::license::*;
-pub use self::openapi::*;
-pub use self::operation::*;
-pub use self::parameter::*;
-pub use self::paths::*;
-pub use self::reference::*;
-pub use self::schema::*;
-pub use self::security::*;
-pub use self::tag::*;
-pub use self::util::*;
+pub use self::{
+    autorest::*, contact::*, external_documentation::*, header::*, info::*, license::*, openapi::*, operation::*, parameter::*, paths::*,
+    reference::*, schema::*, security::*, tag::*, util::*,
+};